requirements.txt
Flask==2.3.3
Pillow>=10.1.0
qrcode==7.4.2
numpy>=1.26.0
setuptools>=65.5.1
wheel>=0.38.0
